Refactor the Rust code
* Clean up mod.rs files
* Format every file
