+++ cat-launcher/src-tauri/src/infra/http_client.rs
@@ -1,9 +1,10 @@
-use reqwest::Client;
 use std::sync::LazyLock;
 
+use reqwest::Client;
+
 pub static HTTP_CLIENT: LazyLock<Client> = LazyLock::new(|| {
-	Client::builder()
-		.user_agent("cat-launcher")
-		.build()
-		.expect("Failed to build reqwest client")
+    Client::builder()
+        .user_agent("cat-launcher")
+        .build()
+        .expect("Failed to build reqwest client")
 });
