+++ cat-launcher/src-tauri/src/infra/mod.rs
@@ -1,3 +1,3 @@
 pub mod http_client;
-pub mod utils;
 pub mod rfc3339;
+pub mod utils;
