+++ cat-launcher/src-tauri/src/lib.rs
@@ -1,7 +1,7 @@
 mod basic_info;
-mod variants;
 mod fetch_releases;
 mod infra;
+mod variants;
 
 use crate::basic_info::commands::get_game_variants_info;
 use crate::fetch_releases::commands::fetch_releases_for_variant;
@@ -10,7 +10,10 @@ use crate::fetch_releases::commands::fetch_releases_for_variant;
 pub fn run() {
     tauri::Builder::default()
         .plugin(tauri_plugin_opener::init())
-        .invoke_handler(tauri::generate_handler![get_game_variants_info, fetch_releases_for_variant])
+        .invoke_handler(tauri::generate_handler![
+            get_game_variants_info,
+            fetch_releases_for_variant
+        ])
         .run(tauri::generate_context!())
         .expect("error while running tauri application");
 }
